fix smoke tests (#809)
* fix smoke tests

* change variable to lower case and add log statement to generateNPI

* oops, generate correct NPI

* renamed NPIValidationUtil to NPIUtil and moved generate NPI to NPIUtil
